climate_guard/requirements.txt
django==4.2.20
djangorestframework==3.14.0
python-dotenv==1.0.0
channels==4.0.0
channels-redis==4.1.0
redis==4.6.0
celery==5.3.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
xgboost==1.7.6
pillow==10.0.0
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
requests-cache==1.1.0
django-cors-headers==4.2.0
